+++ modules/actions-v2/update-user/update-user.php
@@ -411,7 +411,7 @@ class Update_User implements Action_Integration_Interface {
 		$legacy_suffix = ':' . self::SCAN_SCHEMA_VERSION;
 
 		return strlen( $dismissed ) > strlen( $legacy_suffix )
-			&& $legacy_suffix === substr( $dismissed, -strlen( $legacy_suffix ) );
+			&& substr( $dismissed, -strlen( $legacy_suffix ) ) === $legacy_suffix;
 	}
 
 	private function get_notice_dismiss_token( array $notice ): string {
