+++ modules/gateways/secure-price-notice.php
@@ -343,7 +343,7 @@ class Secure_Price_Notice {
 		$legacy_suffix = ':' . self::SCAN_SCHEMA_VERSION;
 
 		return strlen( $dismissed ) > strlen( $legacy_suffix )
-			&& $legacy_suffix === substr( $dismissed, -strlen( $legacy_suffix ) );
+			&& substr( $dismissed, -strlen( $legacy_suffix ) ) === $legacy_suffix;
 	}
 
 	private function get_notice_dismiss_token( array $notice ): string {
